docs/wms que falhou/wms-pro/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
alembic==1.14.0
pydantic==2.10.4
pydantic-settings==2.7.0
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1
python-multipart==0.0.12
asyncpg==0.30.0
aiosqlite==0.20.0
jinja2==3.1.4
httpx==0.28.1
pyinstaller==6.13.0
watchdog==4.0.1
